+++ src/Models/ChunkedUpload.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace NETipar\Chunky\Models;
 
+use Carbon\CarbonInterface;
 use Illuminate\Database\Eloquent\Concerns\HasUlids;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Support\Carbon;
 use NETipar\Chunky\Enums\UploadStatus;
 
 /**
@@ -26,10 +26,10 @@ use NETipar\Chunky\Enums\UploadStatus;
  * @property ?string $final_path
  * @property ?array<string, mixed> $metadata
  * @property UploadStatus $status
- * @property ?Carbon $completed_at
- * @property Carbon $expires_at
- * @property Carbon $created_at
- * @property Carbon $updated_at
+ * @property ?CarbonInterface $completed_at
+ * @property CarbonInterface $expires_at
+ * @property CarbonInterface $created_at
+ * @property CarbonInterface $updated_at
  */
 class ChunkedUpload extends Model
 {